rpc/client: Fail fast on any non-temporary dial error
Previously, RPC client unconditionally blocked by context passed into
`WithContext` (e.g. for 15s). In particular, it definitely stuck on
non-temporary error encounter. An example of such errors could be an
incorrect network address or connection refusal.

Now client instantly fails with any irreparable error in which it is
pointless to wait for a change in the connection state. This is achieved
with `grpcstd.FailOnNonTempDialError(true)` option.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>